requirements-pip.txt
numpy==1.13.3
scipy==1.0.0
pandas==0.23.4
matplotlib==2.2.3
moviepy>=0.2.3.2
Pillow>=3.1.1
lmdb>=0.93
opencv-python>=3.4.0.12
cryptography>=2.1.4
h5py>=2.7.1
jupyter>=1.0.0
pymongo>=3.7.1
requests>=2.19.1
six>=1.11.0
tqdm==4.24.0
parsedatetime==2.4
gunicorn==19.9.0
gevent==1.3.5
pyyaml==3.13
flask==1.0.2
flask-cors==3.0.7
